- added phone app - from webapp to PSTN number works - from webapp to PSTN number doesnt work but able to receive on webapp Reason for not working When we call to the DID number the call comes on the web app and we are  able to accept it.
Once call is accepted in the webapp it takes 20-40 seconds for gathering
 IceCandidates process and for the PSTN phone to update that the call
 has been picked up. During these 20-40 seconds the Caller PSTN phone
 doesn't play any ringtone or anything.

And once the PSTN phone also updates that the call has been picked up.
There is no audio for both the parties involved.

This same issue happens on both the web apps your and the one I have built.